Final Submission for Arduino Afternoon (June 10, 2016)
It's the Final Submission for Arduino Afternoon (June 10, 2016). It
contains the fritzing circuit file as well as a  jpg image file. Pin
configuration is mentioned in code and is drawn in fritzing for only one
7 Segment Display. Full and final outcome of this DIY will be a Clock
displaying HH:MM:SS, along with an understanding of "how LED works",
"how code is working to display time" etc.